Fixes give loadout issue
Seems like player.CurrentLoadout does not update?
Players already have a throwable, it's not null, still it gives them the
loadout over and over again, forcing them back to C4.

Instead it uses a flag to check if loadout was given.

Also tried to move so that primary weapon is given last to force primary
weapon into players hands but this does not seem to have worked
and for some reason it still forces the C4 into their hands.